+++ src/containers/occupation/OccupationListContainer.js
@@ -2,61 +2,51 @@ import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 import OccupationListForm from '../../components/occupation/OccupationListForm';
-import { getOccupationList } from '../../modules/occupation/occupationList';
-import { checkExpire } from '../../lib/api/common/authAPI';
-import { logout } from '../../modules/common/auth';
 import {
   changeInput,
-  postOccupation,
   initializeForm,
-} from '../../modules/occupation/occupationPost';
-
-import {
-  changeInputUpdate,
   initializeResult,
+  getOccupationList,
+  postOccupation,
   updateOccupation,
-} from '../../modules/occupation/occupationUpdate';
-
-import { deleteOccupation } from '../../modules/occupation/occupationDelete';
+  deleteOccupation,
+} from '../../modules/occupation/occupationList';
+import { checkExpire } from '../../lib/api/common/authAPI';
+import { logout } from '../../modules/common/auth';
+import createRequestSaga from '../../lib/createRequestSaga';
 
 const OccupationListContainer = ({ history }) => {
   const [error, setError] = useState(null);
 
   const dispatch = useDispatch();
   const {
-    occupationPost,
-    occupationPostResult,
-    occupationPostError,
-    post,
-    occupationUpdate,
-    occupationUpdateResult,
-    occupationUpdateError,
-    occupationDelete,
-    user,
     occupations,
+    occupationResult,
     occupationError,
     loading,
+    user,
+    occupationPost,
+    occupationUpdate,
+    occupationDelete,
     selectedBranch,
   } = useSelector(
     ({
       occupationPost,
-      auth,
       occupationUpdate,
       occupationDelete,
       occupationList,
+      auth,
       loading,
       select,
     }) => ({
       occupations: occupationList.occupations,
       occupationError: occupationList.occupationError,
-      occupationPost: occupationPost.post,
-      occupationPostResult: occupationPost.occupationPostResult,
-      occupationPostError: occupationPost.occupationPostError,
+      occupationResult: occupationList.occupationResult,
+      occupationPost: occupationList.post,
+      occupationUpdate: occupationList.occupationUpdate,
+      deleteOccupation: occupationList.deleteOccupation,
+      occupationDelete: occupationList.occupationDelete,
       user: auth.user,
-      post: occupationUpdate.post,
-      occupationUpdate: occupationUpdate.occupationUpdate,
-      occupationUpdateResult: occupationUpdate.occupationUpdateResult,
-      occupationUpdateError: occupationUpdate.occupationUpdateError,
       loading: loading,
       selectedBranch: select.selectedBranch,
     }),
@@ -74,8 +64,7 @@ const OccupationListContainer = ({ history }) => {
 
   const onSubmit = (e) => {
     e.preventDefault();
-
-    const data = post;
+    const data = occupationPost;
     if ([data.name, data.color].includes('')) {
       setError('빈 칸을 모두 입력하세요');
       return;
@@ -92,24 +81,13 @@ const OccupationListContainer = ({ history }) => {
     );
   };
 
-  const onEdit = (e) => {
-    e.preventDefault();
-    const data = occupationUpdate;
-    if ([data.name, data.color].includes('')) {
-      setError('빈 칸을 모두 입력하세요');
-      return;
-    }
-
-    // const idx = occupations.map((occupation, index) => <li key={index} />);
-    // dispatch(occupationUpdate({ index, data }));
+  const onDelete = () => {
+    const occupation = occupations.map((occupation, index) => (
+      <li key={index}>{occupation.name}</li>
+    ));
+    console.log(occupation.name);
+    dispatch(deleteOccupation());
   };
-
-  //  const onDelete = (id) => {
-  //    const result = occupations.filter((test) => test.id !== id);
-  //    console.log(test.post.color);
-  //    console.log(test.post.name);
-  //     dispatch(occupationDelete({}));
-  //  };
 
   useEffect(() => {
     if (user !== null) {
@@ -123,31 +101,11 @@ const OccupationListContainer = ({ history }) => {
   }, [dispatch, selectedBranch, user]);
 
   useEffect(() => {
-    if (occupationPostResult === 200) {
+    if (occupationResult === 200) {
       dispatch(initializeForm());
       dispatch(getOccupationList({ selectedBranch }));
     }
-  }, [occupationPostResult, dispatch, selectedBranch]);
-
-  useEffect(() => {
-    if (occupationUpdateResult === '200') {
-      // TODO   리 랜더링 하기
-      dispatch(initializeResult());
-      history.push('/occupation');
-    }
-  }, [occupationUpdateResult, history, dispatch]);
-
-  useEffect(() => {
-    if (occupationPostError !== null) {
-      setError('등록에 실패 했습니다');
-    }
-  }, [occupationPostError]);
-
-  useEffect(() => {
-    if (occupationUpdateError !== null) {
-      setError('수정에 실패 했습니다');
-    }
-  }, [occupationUpdateError]);
+  }, [occupationResult, dispatch, selectedBranch]);
 
   return (
     <OccupationListForm
@@ -156,8 +114,8 @@ const OccupationListContainer = ({ history }) => {
       loading={loading}
       onSubmit={onSubmit}
       onChange={onChange}
-      onEdit={onEdit}
-      // onDelete={onDelete}
+      //onEdit={onEdit}
+      onDelete={onDelete}
       error={error}
     ></OccupationListForm>
   );
