+++ src/lib/api/occupation/occupationAPI.js
@@ -16,6 +16,7 @@ export const updateOccupation = async ({ no, data }) => {
 };
 
 export const deleteOccupation = async ({ no }) => {
-  const response = await client.delete(`/occupation/${no}`);
+  // const response = await client.delete(`/occupation/${no}`);
+  const response = await client.delete(`/occupation/13`);
   return response;
 };
